apply banyak ampe bgnung

+++ static/js/script.js
@@ -1,8 +1,7 @@
-var numOfCard = 8;
+var numOfCard;
 var numOfCardOpened = 0;
 var noRepeatArray = [];
 var finalUrl;
-var level = 1;
 
 function finishRedirect() {	
 	updateSession("gameStateOnGoing", "No");
@@ -10,7 +9,7 @@ function finishRedirect() {
 	window.location = 'feedback';
 }
 function alertPopUp() {	
-	alert("The game has finished! You will be directed to the feedback page in 3 seconds");
+	alert("You've finished the game! You will be directed to the feedback page in 3 seconds");
 }
 
 function shuffle(array) {
@@ -46,13 +45,92 @@ function updateSession(key, value) {
 
 $(document).ready(() => {
 	
-	$('#playNowButton').click(function(e) {
+	$('.readMore').click(function(e) {
+		if ($('.dots').css('display') === "none") {
+			$('.dots').css('display', 'inline');
+			$('.readMore').text('Read more');
+			$('.more').css('display', 'none');
+		} else {
+			$('.dots').css('display', 'none');
+			$('.readMore').text('Read less');
+			$('.more').css('display', 'inline');
+		}
+	})
+
+	$('.expand').click(function(e) {
+		var mamake = $(this).parents("div.item-details.text-center");
+		var dots = $(this).parents("div.item-details.text-center").children("span.dots2");
+		var more = $(this).parents("div.item-details.text-center").children("span.more2")
+
+		if (dots.css('display') === "none") {
+			dots.css('display', 'inline');
+			$(this).text('▼');
+			more.css('display', 'none');
+		} else {
+			dots.css('display', 'none');
+			$(this).text('▲');
+			more.css('display', 'inline');
+		}
+
+	})
+	
+
+	$('.playNowButton').click(function(e) {
 		window.location.pathname = '/play/'
 	})
 	
 	$('#logoHome').click(function(e) {
 		window.location.pathname = '/'
 	})
+
+	$('#customers-testimonials').owlCarousel( {
+		loop: false,
+		
+		center: false,
+		items: 3,
+		margin: 30,
+		autoplay: true,
+		dots:true,
+    	nav:true,
+		autoplayTimeout: 8500,
+		smartSpeed: 450,
+  		navText: ['<i class="fa fa-angle-left"></i>','<i class="fa fa-angle-right"></i>'],
+		responsive: {
+			0: {
+				items: 1
+			},
+			768: {
+				items: 2
+			},
+			1170: {
+				items: 3
+			}
+		}
+	});
+
+	$('#customers-testimonials2').owlCarousel( {
+		loop: true,
+		center: true,
+		items: 3,
+		margin: 10,
+		autoplay: true,
+		dots:true,
+    	nav:true,
+		autoplayTimeout: 8500,
+		smartSpeed: 450,
+  		navText: ['<i class="fa fa-angle-left"></i>','<i class="fa fa-angle-right"></i>'],
+		responsive: {
+			0: {
+				items: 1
+			},
+			768: {
+				items: 2
+			},
+			1170: {
+				items: 3
+			}
+		}
+	});
 
 	$('#shownCard').tilt({
 		scale: 1.02,	
@@ -63,6 +141,20 @@ $(document).ready(() => {
 		glare: true,
 		maxGlare: .11
 	})
+
+	$('.img-responsive').tilt({
+		scale: 1.2,
+		glare: true,
+		maxGlare: .11
+	})
+
+	$('.img-responsive').hover(function() {
+		$(this).css('border-radius', '10px');
+		}, function() { 
+			$(this).css('border-radius', '0')
+		});
+
+	
   
 	var modal = document.getElementById("myModal");
 	$('.close').click(function(event) {
@@ -97,9 +189,13 @@ $(document).ready(() => {
 
 		modal.style.display = "block";
 		numOfCardOpened++;
+		if (sessionStorage.getItem('Players') == 'Two') {
+			player = (numOfCardOpened % 2 != 0) ? sessionStorage.getItem("firstPlayer") : sessionStorage.getItem("secondPlayer");
+			$('#playerTurn').text(player);
+		} else if (sessionStorage.getItem('Players') == 'Multi'){
+			$('#multiHide').css('display', 'none');
+		}
 
-		player = (numOfCardOpened % 2 != 0) ? sessionStorage.getItem("firstPlayer") : sessionStorage.getItem("secondPlayer");
-		$('#playerTurn').text(player);
 
 		$.ajax({
 			method: 'GET',
@@ -143,9 +239,6 @@ $(document).ready(() => {
 			$('#LoveBirdsOption').removeAttr('disabled');
 			$('#LoveBirdsOption').css('color', '#495057');
 
-			$('#MixAndMatchOption').removeAttr('disabled');
-			$('#MixAndMatchOption').css('color', '#495057');
-
 		} else {
 			$('.playerName').removeClass('show')
 
@@ -154,9 +247,6 @@ $(document).ready(() => {
 
 			$('#LoveBirdsOption').attr('disabled', 'disabled');
 			$('#LoveBirdsOption').css('color', '#d3d3d3');
-
-			$('#MixAndMatchOption').attr('disabled', 'disabled');
-			$('#MixAndMatchOption').css('color', '#d3d3d3');
 		}
   	});
 
@@ -181,21 +271,36 @@ $(document).ready(() => {
 		sessionStorage.setItem("gameStateOnGoing", "Yes")
 
 		var pack = $('#pack').val();
-		
+		var level = $('#levelSelection').val();
+		numOfCard = $('#numOfCardSelection').val();
+		sessionStorage.setItem("numOfCard", numOfCard)
+
 		// sessionStorage.setItem("pack", NAMA MODEL DI DJANGOMODEL)
 		if (pack == "Dig Deeper") {
-			sessionStorage.setItem("pack", "DigDeeper")
+			
+			if (level == 3) {
+				sessionStorage.setItem("pack", "DigDeeper3")
+			} else if (level == 2) {
+				sessionStorage.setItem("pack", "DigDeeper2")
+			} else if (level == 1) {
+				sessionStorage.setItem("pack", "DigDeeper1")
+			}
+			
 		} else if (pack == "Break the Ice: Let the Fun Begin!") {
 			sessionStorage.setItem("pack", "IceBreak")
 		} else if (pack == "Mix & Match") {
 
 		} else if (pack == "Lovebirds Edition") {
 
-		} else if (pack == "Hello, Nice to Meet You, Stranger") {
+		} else if (pack == "Hello Nice to Meet You, Stranger") {
 
+		} else if (pack == "Perspective") {
+			sessionStorage.setItem("pack", "Perspective")
 		}
+		
 
 		if ($('#numOfPlayer').val() == 'Two Players') {
+			sessionStorage.setItem('Players', 'Two')
 			$('.playerName').toggleClass('.show')
 			
 			var firstPlayer = $('#firstPlayer').val() != "" ? sessionStorage.setItem("firstPlayer", $('#firstPlayer').val()) : sessionStorage.setItem("firstPlayer", "Player 1")
@@ -204,6 +309,7 @@ $(document).ready(() => {
 			finalUrl = 'form?' + 'pack=' + pack + '&firstPlayer=' + firstPlayer + '&secondPlayer=' + secondPlayer;   
 
 		} else {
+			sessionStorage.setItem('Players', 'Multi')
 			finalUrl = 'form?' + 'pack=' + pack;
 		}
 
@@ -217,16 +323,33 @@ $(document).ready(() => {
 			},
 			success: function(response) {
 				window.location = '/play'
+				
 			}
 		}) 
   	})
-	  
+
+	$('#showCard').click(function (event) {
+		var numOfCard = sessionStorage.getItem("numOfCard")
+		if (numOfCard == 12) {
+			$("#gamePlace12").toggleClass("hidden");
+		} else if (numOfCard == 16) {
+			$("#gamePlace16").toggleClass("hidden");
+		} else if (numOfCard == 20) {
+			$("#gamePlace20").toggleClass("hidden");
+		}
+	})
+
+	
 
 	$('#finishButton').click(function (event) {
-		event.preventDefault()
-		updateSession("gameStateOnGoing", "No")
-		sessionStorage.clear()
-		window.location = '/'
+		event.preventDefault();
+		var alertPopUp = confirm("Oops! You haven't picked all the cards yet. Are you sure you want to finish the game?");
+	
+			if (alertPopUp == true) {
+				updateSession("gameStateOnGoing", "No")
+				sessionStorage.clear()
+				window.location = '/'
+			}
 	})
 
 	$('#terminateLink').click(function (event) {
@@ -239,7 +362,7 @@ $(document).ready(() => {
 		if (sessionStorage.getItem("gameStateOnGoing") != undefined) {
 
 			if (sessionStorage.getItem("gameStateOnGoing") === "Yes" && window.location.pathname === '/play/') {
-				var alertPopUp = confirm("Oops! You have a game going on. Do you wish to terminate the game?");
+				var alertPopUp = confirm("Oops! You still have a game going on. Do you wish to terminate the game?");
 	
 				if (alertPopUp == true) {
 					event.preventDefault();
@@ -247,7 +370,7 @@ $(document).ready(() => {
 					sessionStorage.clear()
 					
 					if ($(this).attr('id') == "playNavBar") {
-						window.location = '/';
+						window.location = 'form';
 					} else {
 						window.location = $(this).attr('href');
 					}
@@ -261,6 +384,11 @@ $(document).ready(() => {
 			window.location = $(this).attr('href');
 		}
 
+	})
+
+	$('.link').click(function (e){
+		e.preventDefault();
+		window.open($(this).attr('href'), '_blank')
 	})
 
 	// $(window).on("beforeunload", function(e) {
